Create getAllMeals.js module
