+++ src/index.html
@@ -25,6 +25,6 @@
 
     <footer>Created by Microverse under CC License.</footer>
     
-    <script src="./index.js"></script>
+    <script type="module" src="./index.js"></script>
   </body>
 </html>
